server/monitoring/scheduler.py
"""
Performance Monitoring Scheduler (Phase 2)

Background scheduler for automated snapshot collection.

PHASE 2 - NOT YET IMPLEMENTED
This is a placeholder/design document for Phase 2 functionality.

To implement Phase 2:
1. Install APScheduler: Add 'apscheduler' to requirements.txt
2. Uncomment the code below
3. Set performance_monitoring.scheduled_snapshots.enabled = true in settings.yaml
4. Configure interval_minutes for health and query snapshots
5. Initialize scheduler in server.py startup
6. Restart the MCP server

Security: All scheduled queries are READ ONLY, monitored by existing security layers
"""

# Phase 2 implementation - currently commented out
# from apscheduler.schedulers.background import BackgroundScheduler
from datetime import datetime
import logging

logger = logging.getLogger(__name__)


# Phase 2 - Placeholder class (not functional until APScheduler is installed)
class PerformanceScheduler:
    """
    Background scheduler for automated performance snapshots
    
    PHASE 2 ONLY - This is a stub implementation.
    Uncomment the actual implementation below after installing APScheduler.
    """
    
    def __init__(self, enabled: bool = False):
        logger.info("Performance monitoring scheduler stub loaded (Phase 2 - not implemented)")
        self.enabled = False
        self.scheduler = None
    
    def add_health_job(self, func, interval_minutes: int, args=None):
        logger.debug("Scheduler stub - health job not added (Phase 2)")
    
    def add_query_job(self, func, interval_minutes: int, args=None):
        logger.debug("Scheduler stub - query job not added (Phase 2)")
    
    def start(self):
        logger.info("Scheduler stub - not starting (Phase 2)")
    
    def shutdown(self):
        logger.info("Scheduler stub - nothing to shutdown (Phase 2)")


"""
# ============================================================================
# PHASE 2 IMPLEMENTATION - Uncomment after installing APScheduler
# ============================================================================

from apscheduler.schedulers.background import BackgroundScheduler

class PerformanceScheduler:
    '''Background scheduler for automated performance snapshots'''
    
    def __init__(self, enabled: bool = False):
        '''
        Initialize scheduler
        
        Args:
            enabled: Whether to start scheduler (default False - Phase 2)
        '''
        self.enabled = enabled
        self.scheduler = None
        
        if self.enabled:
            logger.info("Performance monitoring scheduler ENABLED")
            self.scheduler = BackgroundScheduler()
        else:
            logger.info("Performance monitoring scheduler DISABLED (Phase 2 - set enabled=true to activate)")
    
    def add_health_job(self, func, interval_minutes: int, args=None):
        '''
        Schedule system health collection
        
        Args:
            func: Function to call for health collection
            interval_minutes: Collection frequency
            args: Arguments to pass to func
        '''
        if not self.enabled:
            logger.debug("Scheduler disabled - health job not added")
            return
        
        job_id = f"health_collection_{interval_minutes}min"
        self.scheduler.add_job(
            func,
            'interval',
            minutes=interval_minutes,
            args=args or [],
            id=job_id,
            replace_existing=True
        )
        logger.info(f"Added health collection job: every {interval_minutes} minutes")
    
    def add_query_job(self, func, interval_minutes: int, args=None):
        '''
        Schedule top queries collection
        
        Args:
            func: Function to call for query collection
            interval_minutes: Collection frequency
            args: Arguments to pass to func
        '''
        if not self.enabled:
            logger.debug("Scheduler disabled - query job not added")
            return
        
        job_id = f"query_collection_{interval_minutes}min"
        self.scheduler.add_job(
            func,
            'interval',
            minutes=interval_minutes,
            args=args or [],
            id=job_id,
            replace_existing=True
        )
        logger.info(f"Added query collection job: every {interval_minutes} minutes")
    
    def start(self):
        '''Start the scheduler'''
        if self.enabled and self.scheduler:
            self.scheduler.start()
            logger.info("Performance monitoring scheduler STARTED")
        else:
            logger.info("Scheduler not started (disabled in settings)")
    
    def shutdown(self):
        '''Shutdown the scheduler'''
        if self.scheduler:
            self.scheduler.shutdown()
            logger.info("Performance monitoring scheduler STOPPED")


# ============================================================================
# Example Usage for Phase 2 Implementation
# ============================================================================

Example integration in server.py:

from config import config
from monitoring import OracleMonitor, SnapshotManager
from monitoring.scheduler import PerformanceScheduler

# Check if scheduler is enabled
scheduler_config = config.performance_monitoring.get('scheduled_snapshots', {})
scheduler_enabled = scheduler_config.get('enabled', False)

if scheduler_enabled:
    scheduler = PerformanceScheduler(enabled=True)
    
    # Add jobs for each database with monitoring enabled
    for db_name, db_config in config.database_presets.items():
        monitoring = db_config.get('performance_monitoring', {})
        if monitoring.get('enabled', False):
            
            # System health collection
            if monitoring.get('allow_system_stats', False):
                def collect_health(db_name=db_name):
                    with oracle_connector(db_name) as conn:
                        monitor = OracleMonitor(conn)
                        health = monitor.get_system_health()
                        snapshot_mgr = SnapshotManager()
                        snapshot_mgr.save_health_snapshot(db_name, health)
                
                scheduler.add_health_job(
                    collect_health, 
                    interval_minutes=scheduler_config.get('system_health_interval_minutes', 5)
                )
            
            # Top queries collection
            if monitoring.get('allow_top_queries', False):
                def collect_queries(db_name=db_name):
                    with oracle_connector(db_name) as conn:
                        monitor = OracleMonitor(conn)
                        queries = monitor.get_top_queries_realtime('cpu', 60, 10)
                        snapshot_mgr = SnapshotManager()
                        snapshot_mgr.save_query_snapshots(
                            db_name, 
                            datetime.now(), 
                            queries.get('queries', []),
                            'cpu'
                        )
                
                scheduler.add_query_job(
                    collect_queries,
                    interval_minutes=scheduler_config.get('top_queries_interval_minutes', 15)
                )
    
    scheduler.start()
    logger.info("✅ Phase 2 scheduler started with background collection")

"""